Add functionality for getting all branches
This patch adds work to get all of the branches of a repository. It
does this via passing "this" as the payload which is then used
to call the branches function, which populates a temporary variable
repoBranches. The temporary variable is then looped over to get the
branch names which is passed back to the user. The temp variable is then
cleared.